Replace requests lib with aiohttp for non-blocking appointment fetching
The function that fetches appointments was blocking the main thread, mainly
because it was using the `requests` library that does not support
async/await.

That meant that during the `get_appointment` duration (which effectively
was the latency of 2 requests + 1 second of sleep) the websocket server
was not able to attend new connections.

Moving to the `aiohttp` library freed the main thread to attend new
incoming connections.